Use treehouse project as starting point.
Replaced express starter app with treehouse tutorial as starting point.
Removed eventonica references (we'll get to that next week. I'm sorry if the intention was to start it this assessment!)
Postpone migrations, testing. 
Add a note that we'll cover REST soon. 
Remove references to MySQL, because we're using Postgres this cohort.