PR on behalf of @infiniteverma: Format Slideshows Part 1/3
solves #598.  I have not heard from this person in weeks, but it seems finished, and I think they should get the points.

Mentors, please choose 5 at random and make sure it follows the format mentioned in #598.  If it doesn't, please comment about it.  Slides 30 is an exception, because neither @InfiniteVerma nor I could get edit rights to it.