Update enzyme-testing with Testing Automation link.md (#1223)
Thank you @abhinashgiri  ! In future commit messages, be sure to be specific about what sort of update you are making (ex: "Update links in __ section of enzyme-testing.md" or "Fix wording in motivation section". That way, if you have to go back to a certain version, it will be easier to find by looking at the commit messages.