Update react-part-1-intro.md
There are a few things that will need to be updated over time. In the React slide deck, it stated that the react-create-class module did not need to be learned. That conflicted with having a tutorial use that module. Also, many of the tutorials are all done with classes instead of functions. This will be a task that I could fix in a few different steps but to start, I updated the slide and added #4 to the Common Mistakes/ Misconceptions to shed light on defining components with functions in React.