Restructured companies list
Replaced the companies list with a table to hopefully help participants recognize the names of some companies they've heard of and get them familiar with some of the biggest projects. Since they're becoming JS developers, I want to make sure React is near the top, but I'll fill in more shortly. 

I realized as I started this that I should link to the OSPO page for each organization, but I need to explain what that is first, so that'll be a different commit then I'll come back here to fill this out.